requirements.txt
numpy
pillow
onnxruntime
python-telegram-bot